+++ packages/expo-ui/android/src/main/java/expo/modules/ui/ToggleButtonView.kt
@@ -16,6 +16,7 @@ import expo.modules.kotlin.records.Record
 import expo.modules.kotlin.views.ComposeProps
 import expo.modules.kotlin.views.FunctionalComposableScope
 import expo.modules.kotlin.types.OptimizedRecord
+import expo.modules.kotlin.views.OptimizedComposeProps
 
 @OptimizedRecord
 data class ToggleButtonColors(
@@ -27,6 +28,7 @@ data class ToggleButtonColors(
   @Field val disabledContentColor: Color? = null
 ) : Record
 
+@OptimizedComposeProps
 data class ToggleButtonProps(
   val checked: Boolean = false,
   val enabled: Boolean = true,