+++ packages/expo-ui/android/src/main/java/expo/modules/ui/menu/DropdownMenuRecords.kt
@@ -4,12 +4,14 @@ import android.graphics.Color
 import expo.modules.kotlin.types.Enumerable
 import expo.modules.kotlin.views.ComposeProps
 import expo.modules.ui.ModifierList
+import expo.modules.kotlin.views.OptimizedComposeProps
 
 enum class ActivationMethod(val value: String) : Enumerable {
   SINGLE_PRESS("singlePress"),
   LONG_PRESS("longPress")
 }
 
+@OptimizedComposeProps
 data class DropdownMenuProps(
   val activationMethod: ActivationMethod = ActivationMethod.SINGLE_PRESS,
   val expanded: Boolean = false,